requirements.txt
Flask
Werkzeug==2.0.3
Pillow
moviepy
